Refactor:#21 BaseEntity id 타입 변경에 따른 리팩토링

+++ back/src/main/java/com/back/auth/adapter/in/web/dto/AuthMemberResponse.java
@@ -4,7 +4,7 @@ import com.back.member.domain.Member;
 
 /** 인증 API에서 사용하는 회원 정보 응답 DTO. */
 public record AuthMemberResponse(
-    Integer id, String email, String nickname, String role, String status) {
+    Long id, String email, String nickname, String role, String status) {
 
   public static AuthMemberResponse from(Member member) {
     return new AuthMemberResponse(
