+++ back/src/main/java/com/back/member/adapter/in/web/dto/MemberResponse.java
@@ -2,7 +2,7 @@ package com.back.member.adapter.in.web.dto;
 
 import com.back.member.domain.Member;
 
-public record MemberResponse(Integer id, String email, String nickname) {
+public record MemberResponse(Long id, String email, String nickname) {
 
   public static MemberResponse from(Member member) {
     return new MemberResponse(member.getId(), member.getEmail(), member.getNickname());
