+++ back/src/test/java/com/back/member/adapter/in/web/MemberControllerTest.java
@@ -53,7 +53,7 @@ class MemberControllerTest {
   void createMember() throws Exception {
     CreateMemberRequest request =
         new CreateMemberRequest("member1@test.com", "pass1234", "member1");
-    MemberResponse response = new MemberResponse(1, "member1@test.com", "member1");
+    MemberResponse response = new MemberResponse(1L, "member1@test.com", "member1");
 
     given(memberService.createMember(any(CreateMemberRequest.class))).willReturn(response);
 
@@ -73,8 +73,8 @@ class MemberControllerTest {
   @Test
   @DisplayName("관리자는 memberId 기반 회원 조회 API를 사용할 수 있다")
   void getMember() throws Exception {
-    MemberResponse response = new MemberResponse(2, "member2@test.com", "member2");
-    given(memberService.getMember(2)).willReturn(response);
+    MemberResponse response = new MemberResponse(2L, "member2@test.com", "member2");
+    given(memberService.getMember(2L)).willReturn(response);
 
     mockMvc
         .perform(get("/api/v1/members/{memberId}", 2).with(user("admin").roles("ADMIN")))
@@ -90,13 +90,13 @@ class MemberControllerTest {
   @DisplayName("본인 프로필 수정 API는 인증된 사용자 기준으로 닉네임을 변경한다")
   void updateProfile() throws Exception {
     UpdateMemberProfileRequest request = new UpdateMemberProfileRequest("updatedMember");
-    MemberResponse response = new MemberResponse(3, "member3@test.com", "updatedMember");
-    given(memberService.updateProfile(3, request)).willReturn(response);
+    MemberResponse response = new MemberResponse(3L, "member3@test.com", "updatedMember");
+    given(memberService.updateProfile(3L, request)).willReturn(response);
 
     mockMvc
         .perform(
             patch("/api/v1/members/me/profile")
-                .with(authentication(authenticatedMember(3, "member3@test.com")))
+                .with(authentication(authenticatedMember(3L, "member3@test.com")))
                 .contentType(APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(request)))
         .andExpect(status().is2xxSuccessful())
@@ -123,7 +123,7 @@ class MemberControllerTest {
         .andExpect(jsonPath("$.msg").value("You do not have permission."));
   }
 
-  private UsernamePasswordAuthenticationToken authenticatedMember(Integer memberId, String email) {
+  private UsernamePasswordAuthenticationToken authenticatedMember(Long memberId, String email) {
     AuthenticatedMember principal = new AuthenticatedMember(memberId, email, List.of("ROLE_USER"));
     return new UsernamePasswordAuthenticationToken(
         principal, null, List.of(new SimpleGrantedAuthority("ROLE_USER")));
